move all filtering methods/vars from `historyStore/model/filtering` to `utils/filterConversation`
