+++ client/src/utils/filtering.test.js
@@ -1,9 +1,10 @@
-import { checkFilter, getFilters, toAlias, getQueryDict, testFilters } from "./filtering";
+import { checkFilter, getFilters, getQueryDict, testFilters, toAlias } from "utils/filterConversion";
 
 const filterTexts = [
     "name:'name of item' hid>10 hid<100 create-time>'2021-01-01' update-time<'2022-01-01' state:success extension:ext tag:first deleted:False visible:'TRUE'",
     "name:'name of item' hid_gt:10 hid-lt:100 create_time-gt:'2021-01-01' update_time-lt:'2022-01-01' state:sUccEss extension:EXT tag:FirsT deleted:false visible:true",
 ];
+
 describe("filtering", () => {
     test("parse default filter", () => {
         let queryDict = getQueryDict("");
